[systemd] Made .deb package systemd compliant
Without breaking compatibility with SysVInit, all of it in a single
package (maybe in a near future making separate package using runit in
Omnibus would be a better option).

Some cleanup has also been made (replacing which by command -v to be
POSIX compliant, moving a chmod call from the postinst script to the
package extraction phase, adding one or two echoes...)

This is actually a squash of 9 commits, I let the individual messages
below, feel free to have a look.

[systemd] made the agent systemd compliant

Everything is in the title. Careful there's also a modification on
install_agent.sh in the dd-agent project.

[systemd] Restart after install fixed

The postinst script restarts systemd under the same condition as it did
before for SysVInit.

[systemd] which calls replaced by command -v

[systemd] which replaced in pre scripts as well

[systemd] Code cleanup and optimisations

Where it was possible we use update-rc.d (actually everywhere but with
some slight modification to to the fact that systemctl and SysVInit
behave differently, especially when it comes to exit values.

[systemd] Last rectifications before omnibus build

[systemd] last optim.

[systemd] .service file added to soft config

Also an echo has been added to the postinst script for a better user
feedback when using systemd and the --force option is now used when
asking for service removal with update-rc.d.

[systemd] Moved chmod 755 from postinst to .deb

The chmod 755 on /etc/init.d/datadog-agent was in the post-install
script. It should rather be done at the end of the package extraction
(like it's done for all the other files), I put it there.